% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_explore.R
\name{acctionbutton_col}
\alias{acctionbutton_col}
\title{Add Action Button to a Column in a Data Frame}
\usage{
acctionbutton_col(df, col, ns, inputId)
}
\arguments{
\item{df}{A data frame to which the action button will be added.}

\item{col}{A character string specifying the column to which the action
button will be added.}

\item{ns}{A shiny namespace object, usually from `ns <- NS(id)`.}

\item{inputId}{A character string specifying the input ID that will be used
in Shiny's onInputChange function.}
}
\value{
A data frame with the specified column replaced by action buttons.
}
\description{
This function adds an action button to a specified column in a data frame.
The action button is created using Shiny's actionButton function.
}
