% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_explore.R
\name{custom_dt}
\alias{custom_dt}
\title{Create an Interactive DataTable}
\usage{
custom_dt(df)
}
\arguments{
\item{df}{A data frame to be displayed in the DataTable.}
}
\value{
An interactive DataTable.
}
\description{
This function takes a data frame and generates an interactive DataTable
using the DT package with various customization options.
}
\examples{
custom_dt(iris)
}
