% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_home.R
\name{donors_by_exp_plt}
\alias{donors_by_exp_plt}
\title{Create a bar plot showing the number of donors per experiment type.}
\usage{
donors_by_exp_plt(df, source_id)
}
\arguments{
\item{df}{A data frame containing experiment information.}

\item{source_id}{Identifier for the plot's data source.}
}
\value{
A Plotly bar plot visualizing the number of donors per experiment type.
}
\description{
This function takes a data frame and an identifier for the plot's data source
and generates a bar plot using Plotly. The plot displays the number of donors
for each experiment type.
}
