% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivihome-package.R
\docType{package}
\name{ivihome-package}
\alias{ivihome}
\alias{ivihome-package}
\title{ivihome: MISTRAL Data Portal}
\description{
The ivihome (MISTRAL Data Portal) stores and provides HIV realted data generated by MISTRAL Consortium labs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MicrobialGenomics-IrsicaixaOrg/ivihome}
  \item Report bugs at \url{https://github.com/MicrobialGenomics-IrsicaixaOrg/ivihome/issues}
}

}
\author{
\strong{Maintainer}: Francesc Catala-Moll \email{fcatala@irsicaixa.es}

}
\keyword{internal}
