% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistralDP-package.R
\docType{package}
\name{mistralDP-package}
\alias{mistralDP}
\alias{mistralDP-package}
\title{mistralDP: MISTRAL Data Portal}
\description{
The ivihome (MISTRAL Data Portal) stores and provides HIV realted data generated by MISTRAL Consortium labs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MicrobialGenomics-IrsicaixaOrg/mistralDP}
  \item Report bugs at \url{https://github.com/MicrobialGenomics-IrsicaixaOrg/mistralDP/issues}
}

}
\author{
\strong{Maintainer}: Francesc Catala-Moll \email{fcatala@irsicaixa.es}

}
\keyword{internal}
