% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_explore.R
\name{pie_plt_2}
\alias{pie_plt_2}
\title{Generate a Pie Chart}
\usage{
pie_plt_2(
  df,
  group_var = "gender",
  count_var = "patient_id",
  title = NULL,
  ns,
  click_id
)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{group_var}{The name of the categorical variable to group by.}

\item{count_var}{The variable used for counting the occurrences.}

\item{title}{The title of the pie chart.}

\item{ns}{The namespace for Shiny.}

\item{click_id}{Click identification.}
}
\value{
A pie chart created using Plotly.
}
\description{
This function takes a data frame and creates a pie chart to visualize
the distribution of a categorical variable.
}
\examples{
pie_plt(mtcars, "cyl", "mpg", "Cylinder Distribution", "A")
}
