% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_home.R
\name{stat_box}
\alias{stat_box}
\title{stat_box function}
\usage{
stat_box(title, value, icon, icon_size = "4em", opacity = 0.8)
}
\arguments{
\item{title}{The title of the stat box.}

\item{value}{The value displayed in the stat box.}

\item{icon}{The icon to be displayed.}

\item{icon_size}{The size of the icon.}

\item{opacity}{The opacity of the icon.}
}
\value{
A tag containing the styled stat box.
}
\description{
Create a styled stat box with title, value, and icon.
}
