% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_home.R
\name{wp_button}
\alias{wp_button}
\title{Create a customized button.}
\usage{
wp_button(id, lab, ns)
}
\arguments{
\item{id}{The ID for the button input.}

\item{lab}{The label for the button.}

\item{ns}{The namespace for Shiny.}
}
\value{
A Shiny UI element representing the customized button.
}
\description{
This function creates a customized button using the `actionButton()` function
from the Shiny package. The button's appearance is customized through CSS
variables.
}
